util: Fix resource path
Since commit 1f6f1a48 the resource path for icons has been broken.
The reason is that when moving the .ui files to $(srcdir)/resources/ui
the define used for the resources was changed to reflect the new
directory. However, this change wasn't needed by the icons and ended
up with virt-viewer not displaying a few icons.

Let's fix the issue by setting back the define to the previous one and
then tweaking the virt_viewer_util_load_ui() to add "ui" to the resource
path, for loading the ui files.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>