app,vnc: Don't retry to authenticate when vvfile authentication fails
When connecting to a RHEVM/oVirt machine through an expired vvfile the
user ends up in an endless loop of getting an error message, pressing
"Okay", re-scheduling a new connection retry. getting an error message
due to the expired vvfile and so on.

In order to avoid the issue, let's not re-schedule the connection retry
when the user tries to connect through a vvfile and it fails.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>