virt-viewer: fix display of guest name in title for vnc
The following commit broke the display of the guest name in
the title for VNC displays:

  commit 61a1bc4dcbb056755fe96c5945f84c1312041059
  Author: Pavel Grunt <pgrunt@redhat.com>
  Date:   Wed Apr 15 13:50:35 2015 +0200

    session-vnc: Set window for display to avoid gtk-vnc v0.3.8 crash

    The VNC display widget of gtk-vnc v0.3.8 needs a window at the moment

The problem is that this causes the window to be associated
with the display before the guest name is available. Thus
when ensure_window_for_display() runs, the window is already
configured and so it never invokes the logic to set the title.

The fix is to unconditionally update the title in the
ensure_window_for_display() method, even if the window already
exists.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>