vnc: Set display as enabled on init
Since 9c77a78af2ef85f3fcdce21b42d89566a9f7ee17 the vnc display has
stopped setting the show hint and started to ignore the initial zoom
setting. Let's handle it in a similar way as the spice display and set
the hint when the display is initialized and unset it on disconnect.

Resolves: rhbz#1436991

Acked-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>